requirements.txt
#-------------app dependencies-----------------------
python-dotenv~=1.0.1
pydantic~=2.8.2
requests~=2.32.3
openai~=1.44.1
pandas~=2.2.3